# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Audio Player API
  version: 1.0.0
  description: |
    API documentation for Nexus Audio Player - A modern audio/video player with cloud sync capabilities.
    
    ## Authentication
    All protected endpoints require authentication via Bearer token in the Authorization header.
    The token can be either:
    - Firebase ID token (recommended)
    - Google OAuth access token
    
    ## Rate Limiting
    API requests are rate-limited. Contact support if you need higher limits.
    
    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message",
        "details": {}
      }
    }
    ```
  contact:
    name: Nexus Audio Player Support
    email: support@nexus-audio.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.nexus-audio.com/api
    description: Production server
...
paths:
...
  /storage/upload:
    post:
      tags:
        - Storage
      summary: Upload a file
      description: |
        Upload a file to storage. Pro users automatically use Bunny CDN, free users use local storage.
        
        **File size limits:**
        - Free users: 100MB
        - Pro users: 500MB
        
        **Allowed file types:**
        - Audio: MP3, WAV, FLAC, AAC, OGG
        - Video: MP4, AVI, MKV, WebM
        - Images: JPEG, PNG, GIF, WebP
      operationId: uploadFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      examples:
        success:
          summary: Successful upload
          value:
            id: "1234567890-123456789"
            url: "https://cdn.example.com/nexus/user123/file.mp3"
            size: 5242880
            filename: "song.mp3"
            provider: "bunny"
...
  /storage/download/{fileId}:
    get:
      tags:
        - Storage
      summary: Download a file
      description: Download a file by ID (returns file stream)
      operationId: downloadFile
      security:
        - bearerAuth: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
          description: The file ID to download
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /stripe/create-portal-session:
    post:
      tags:
        - Stripe
      summary: Create billing portal session
      description: Create a Stripe billing portal session for managing subscription
      operationId: createPortalSession
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortalRequest'
      responses:
        '200':
          description: Portal session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token authentication. Use either:
        - Firebase ID token (recommended)
        - Google OAuth access token
  schemas:
    UploadResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique file ID
          example: "1234567890-123456789"
        url:
          type: string
          format: uri
          description: URL to access the file
          example: "https://cdn.example.com/nexus/user123/file.mp3"
        size:
          type: integer
          description: File size in bytes
          example: 5242880
        filename:
          type: string
          description: Original filename
          example: "song.mp3"
        provider:
          type: string
          enum: [local, bunny]
          description: Storage provider used
          example: "bunny"
...
    PortalRequest:
      type: object
      properties:
        returnUrl:
          type: string
          format: uri
          description: URL to return to after managing subscription
          example: "https://app.nexus-audio.com/settings"
    PortalResponse:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: Billing portal URL to redirect user to
          example: "https://billing.stripe.com/session/1234567890"
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "File too large. Maximum size is 100MB"
            details:
              type: object
              description: Additional error details
              nullable: true
  responses:
    BadRequest:
      description: Bad request - validation error or invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "File too large. Maximum size is 100MB"
              details:
                maxSize: 104857600
                fileSize: 157286400
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "AUTHENTICATION_ERROR"
              message: "User not authenticated"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "File not found"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
    ServiceUnavailable:
      description: Service unavailable - external service not configured
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "EXTERNAL_SERVICE_ERROR"
              message: "Stripe is not configured. Please set STRIPE_SECRET_KEY in .env"